Some warnings removed
git-svn-id: file:///Volumes/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@15116 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/kernel/core.hpp
@@ -3364,7 +3364,7 @@ namespace Gecode {
   forceinline PropagatorGroup
   ExecInfo::post(void) const {
     assert(what() == POST);
-    return PropagatorGroup(who >> 2);
+    return PropagatorGroup(static_cast<unsigned int>(who >> 2));
   }
 
   /*
